requirements.txt
flet>=0.25.0
google-genai>=0.8.0
